Make the IdGenerator implementation configurable by the user.
