+++ README.md
@@ -18,11 +18,8 @@ TODOs for Release:
 
 - `DynamoDbStoredEventRepository::RetrieveAllAfterVersion()` uses a filter expression which isn't efficient.
 - Handling for manageable DynamoDb errors.
-- A cleaner approach to handling metadata.
-- Ensure package config is correct and install journey is easy and clear.
-- Provide an interface to allow users to replace IdGenerator with their own.
+- Review approach to handling event metadata, ensure its compatible.
 - Copy and modify any parts of the main package test suite that can give more end to end coverage.
-- Write some basic docs. (WIP)
 - See if we can solve for LazyCollection->remember().
 
 ## Features
@@ -33,14 +30,12 @@ TODOs for Release:
 - CreateTables command to get you started quickly.
 - Optional support for [strongly consistent reads](#read-consistency) (with caveats).
 
-## Minor Differences
+### Minor Differences
 
 - The default `EloquentStoredEventRepository:store()` implementation converts a `null` `$uuid` argument to an empty
   string for storage. DynamoDb does not allow empty strings, so we store this as the string `'null'`.
 
 ## How It Works
-
-The gorey DynamoDB details.
 
 ### Events
 
@@ -94,16 +89,32 @@ The gorey DynamoDB details.
 
 ### Event Ids
 
-- This package generates its own Int64 ids for events. This is due to the Spatie package interfaces expecting
-  integer ids and the logic expecting them to be incrementing. Dynamo has no mechanism for returning incrementing ints.
-- The ids consist of the current microsecond timestamp expressed as an integer plus 3 random digits appended to the end.
-- The timestamp component approximates the incrementing id behaviour that's expected for events and the random digits
+- This package generates its own Int64 Ids for events. The Spatie package interfaces expect integer ids and the logic
+  expects them to be incrementing. DynamoDb does not provide incrementing ids.
+- The Ids consist of the current microsecond timestamp expressed as an integer plus 3 random digits appended to the end.
+- The timestamp component approximates the incrementing Id behaviour that's expected for events and the random digits
   increase collision resistance in the unlikely event that two are generated in the same microsecond.
 - Collisions are possible but unlikely. The collision scenario would be two PHP processes generating ids and hitting the
   same microsecond timestamp and then generating the same random 3 digit int to append to them.
 - This collision scenario is unhandled in the code and the consequences would depend on the design of your application.
+- Consider also that, at scale, clock skew between servers could cause issues for this.
 - In the event that the random digits do not clash but the microsecond timestamp is the same, the event ordering
   is determined by the random digits.
+- You can switch to your own implementation for generation of Ids by implementing the `IdGenerator` interface and
+  updating the config key `id_generator`.
+
+```
+    'id_generator' => TimeStampIdGenerator::class,
+```
+
+- You can switch to your own implementation of a timestamp provider for the provided `TimeStampIdGenerator` by
+  implementing the `TimestampProvider` interface and updating the config key `id_timestamp_provider`. If you return a
+  shorter timestamp (e.g. seconds or milliseconds) the IdGenerator will fill the remainder of the 64bit Int with random
+  digits.
+
+```
+    'id_timestamp_provider' => TimeStampIdGenerator::class,
+```
 
 ### LazyCollections
