+++ tests/IdGeneration/IdGeneratorTest.php
@@ -1,7 +1,7 @@
 <?php
 
-use BlackFrog\LaravelEventSourcingDynamodb\IdGeneration\IdGenerator;
 use BlackFrog\LaravelEventSourcingDynamodb\IdGeneration\MicroTimestampProvider;
+use BlackFrog\LaravelEventSourcingDynamodb\IdGeneration\TimeStampIdGenerator;
 use BlackFrog\LaravelEventSourcingDynamodb\IdGeneration\TimestampProvider;
 use Random\Randomizer;
 
@@ -20,7 +20,7 @@ beforeAll(function () {
 });
 
 beforeEach(function () {
-    $this->idGenerator = new IdGenerator(new Randomizer(), new MicroTimestampProvider());
+    $this->idGenerator = new TimeStampIdGenerator(new Randomizer(), new MicroTimestampProvider());
 });
 
 afterEach(function () {
@@ -98,7 +98,7 @@ it('produces no duplicate ids over 50,000 iterations', function () {
 
 it('uses the provided integer timestamp as the start of each id', function () {
     $microTimeStamp = 1679836125000000;
-    $this->idGenerator = new IdGenerator(new Randomizer(), new FixedTimestampProvider($microTimeStamp));
+    $this->idGenerator = new TimeStampIdGenerator(new Randomizer(), new FixedTimestampProvider($microTimeStamp));
 
     $id = $this->idGenerator->generateId();
     $idAsString = (string) $id;
@@ -115,9 +115,9 @@ it('may generate duplicate ids if two instances are used at the same microsecond
 
     //450 iterations in one microsecond guarantees at least some collisions between two instances.
     $iterations = 450;
-    $idGenerator1 = new IdGenerator(new Randomizer(), $timestampProvider);
+    $idGenerator1 = new TimeStampIdGenerator(new Randomizer(), $timestampProvider);
     $idsGenerated1 = [];
-    $idGenerator2 = new IdGenerator(new Randomizer(), $timestampProvider);
+    $idGenerator2 = new TimeStampIdGenerator(new Randomizer(), $timestampProvider);
     $idsGenerated2 = [];
 
     $x = 1;
@@ -135,7 +135,7 @@ it('generates ids even when the system date is a unix date early in the epoch', 
     //Fix the current time to 1 microsecond since the start of unix epoch.
     $microTimeStamp = 1;
     $timestampProvider = new FixedTimestampProvider($microTimeStamp);
-    $this->idGenerator = new IdGenerator(new Randomizer(), $timestampProvider);
+    $this->idGenerator = new TimeStampIdGenerator(new Randomizer(), $timestampProvider);
 
     $id = $this->idGenerator->generateId();
 
@@ -150,7 +150,7 @@ it('generates ids with reduced entropy when system date is 2286-11-20 onwards', 
     //The id generator soldiers on with reduced entropy by generating one less random digit to append.
     $microTimeStamp = 10000000000000000;
     $timestampProvider = new FixedTimestampProvider($microTimeStamp);
-    $this->idGenerator = new IdGenerator(new Randomizer(), $timestampProvider);
+    $this->idGenerator = new TimeStampIdGenerator(new Randomizer(), $timestampProvider);
 
     //The max iterations per microsecond falls to 79 due to fewer available random numbers.
     $iterations = 79;
