+++ tests/Snapshots/DynamoDbSnapshotRepositoryTest.php
@@ -1,8 +1,8 @@
 <?php
 
 use Aws\DynamoDb\Marshaler;
-use BlackFrog\LaravelEventSourcingDynamodb\IdGeneration\IdGenerator;
 use BlackFrog\LaravelEventSourcingDynamodb\IdGeneration\MicroTimestampProvider;
+use BlackFrog\LaravelEventSourcingDynamodb\IdGeneration\TimeStampIdGenerator;
 use BlackFrog\LaravelEventSourcingDynamodb\Snapshots\DynamoDbSnapshotRepository;
 use BlackFrog\LaravelEventSourcingDynamodb\Snapshots\StateSerializer;
 use Random\Randomizer;
@@ -26,7 +26,7 @@ beforeEach(function () {
 
     $this->snapshotRepository = new DynamoDbSnapshotRepository(
         $this->getDynamoDbClient(),
-        new IdGenerator(new Randomizer(), new MicroTimestampProvider()),
+        new TimeStampIdGenerator(new Randomizer(), new MicroTimestampProvider()),
         new Marshaler(),
         new StateSerializer()
     );
